Add locust runner to test local node
